log: add flb_log_suppress_check
