+++ include/fluent-bit/flb_output_plugin.h
@@ -27,36 +27,8 @@
 #include <cmetrics/cmt_encode_text.h>
 #include <cmetrics/cmt_encode_prometheus.h>
 
-#include <stdarg.h>
-
-static inline int flb_output_plugin_log_suppress_check(struct flb_output_instance *ins, const char *fmt, ...)
-{
-    int ret;
-    size_t size;
-    va_list args;
-    char buf[4096];
-    struct flb_worker *w;
-
-    if (ins->log_suppress_interval <= 0) {
-        return FLB_FALSE;
-    }
-
-    va_start(args, fmt);
-    size = vsnprintf(buf, sizeof(buf) - 1, fmt, args);
-    va_end(args);
-
-    if (size == -1) {
-        return FLB_FALSE;
-    }
-
-    w = flb_worker_get();
-    if (!w) {
-        return FLB_FALSE;
-    }
-
-    ret = flb_log_cache_check_suppress(w->log_cache, buf, size);
-    return ret;
-}
+#define flb_output_plugin_log_suppress_check(ins, fmt, ...) \
+    flb_log_suppress_check(ins->log_suppress_interval, fmt, ##__VA_ARGS__)
 
 #define flb_plg_error(ctx, fmt, ...)                                    \
     if (flb_log_check_level(ctx->log_level, FLB_LOG_ERROR))             \
